Disassembled server.js into smaller files.
