+++ server/app.js
@@ -4,7 +4,7 @@
 const Promise = require('bluebird');
 const express = require('express');
 
-const server = require('./server');
+const server = require('./src/server');
 
 const app = express();
 
@@ -20,8 +20,6 @@ var helpersInitialized = new Promise(function(resolve, reject) {
 helpersInitialized.then(function(o) {
   const {
     addCorsHeader,
-    assignToP,
-    assignToPCustom,
     auth,
     authOptional,
     COOKIES,
@@ -34,31 +32,12 @@ helpersInitialized.then(function(o) {
     fetchIndexForConversation,
     fetchIndexForReportPage,
     fetchIndexWithoutPreloadData,
-    getArrayOfInt,
-    // getArrayOfStringLimitLength,
-    getArrayOfStringNonEmpty,
-    getArrayOfStringNonEmptyLimitLength,
-    getBool,
-    getConversationIdFetchZid,
-    getEmail,
-    getInt,
-    getIntInRange,
-    getNumberInRange,
-    getOptionalStringLimitLength,
-    getPassword,
-    getPasswordWithCreatePasswordRules,
     getPidForParticipant,
-    getReportIdFetchRid,
-    getStringLimitLength,
-    getUrlLimitLength,
     haltOnTimeout,
     HMAC_SIGNATURE_PARAM_NAME,
     hostname,
     makeFileFetcher,
     makeRedirectorTo,
-    moveToBody,
-    need,
-    // needHeader,
     pidCache,
     portForAdminFiles,
     portForParticipationFiles,
@@ -67,11 +46,7 @@ helpersInitialized.then(function(o) {
     redirectIfHasZidButNoConversationId,
     redirectIfNotHttps,
     redirectIfWrongDomain,
-    resolve_pidThing,
     timeout,
-    want,
-    wantCookie,
-    wantHeader,
     winston,
     writeDefaultHead,
 
@@ -206,6 +181,34 @@ helpersInitialized.then(function(o) {
     handle_PUT_reports,
     handle_PUT_users,
   } = o;
+
+  const {
+    assignToP,
+    assignToPCustom,
+    getArrayOfInt,
+    getArrayOfStringNonEmpty,
+    getArrayOfStringNonEmptyLimitLength,
+    getBool,
+    getConversationIdFetchZid,
+    getEmail,
+    getInt,
+    getIntInRange,
+    getNumberInRange,
+    getOptionalStringLimitLength,
+    getPassword,
+    getPasswordWithCreatePasswordRules,
+    getReportIdFetchRid,
+    getStringLimitLength,
+    getUrlLimitLength,
+    moveToBody,
+    need,
+    needCookie,
+    needHeader,
+    resolve_pidThing,
+    want,
+    wantCookie,
+    wantHeader,
+  } = require('./src/utils/parameter');
 
   console.log('begin route config');
