+++ server/src/config.js
@@ -1,0 +1,14 @@
+const devMode = require('boolean')(get('DEV_MODE'));
+
+function get(key) {
+	return process.env[key];
+}
+
+function isDevMode() {
+	return devMode;
+}
+
+module.exports = {
+	get,
+	isDevMode
+};
