+++ server/src/conversation.js
@@ -1,0 +1,34 @@
+const pg = require('./db/pg-query');
+const MPromise = require('./utils/metered').MPromise;
+const LruCache = require("lru-cache");
+
+const conversationIdToZidCache = new LruCache({
+  max: 1000,
+});
+
+// NOTE: currently conversation_id is stored as zinvite
+function getZidFromConversationId(conversation_id) {
+  return new MPromise("getZidFromConversationId", function(resolve, reject) {
+    let cachedZid = conversationIdToZidCache.get(conversation_id);
+    if (cachedZid) {
+      resolve(cachedZid);
+      return;
+    }
+    pg.pgQuery_readOnly("select zid from zinvites where zinvite = ($1);", [conversation_id], function(err, results) {
+      if (err) {
+        return reject(err);
+      } else if (!results || !results.rows || !results.rows.length) {
+        console.error("polis_err_fetching_zid_for_conversation_id " + conversation_id);
+        return reject("polis_err_fetching_zid_for_conversation_id");
+      } else {
+        let zid = results.rows[0].zid;
+        conversationIdToZidCache.set(conversation_id, zid);
+        return resolve(zid);
+      }
+    });
+  });
+}
+
+module.exports = {
+  getZidFromConversationId,
+};
